third_party/blink/web_tests/external/wpt/cookiestore/cookieListItem_attributes.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] CookieListItem - cookieStore.set does not add / to path if it does not end with /
  assert_equals: expected "/cookiestore" but got "/cookiestore/"
Harness: the test ran to completion.
